## Before this PR
Similar to #864 , JFRs showing significant allocations and CPU consumed due to `com.palantir.conjure.java.api.config.service.UserAgents#isValidName` and `com.palantir.conjure.java.api.config.service.UserAgents#format` during Dialogue request processing

![image](https://user-images.githubusercontent.com/54594/180318740-42ca6b3b-e678-4d1e-af0c-3b1af9d6931a.png)
![image](https://user-images.githubusercontent.com/54594/180319030-36eb3c1f-0946-4d68-aa52-72500d72154b.png)



## After this PR
Some initial benchmark results on 16 core `Intel(R) Xeon(R) Platinum 8259CL CPU @ 2.50GHz` where `regex` is the previous implementation and `parser` is our hand rolled non-allocating version.

```
JDK 11.0.13, OpenJDK 64-Bit Server VM, 11.0.13+8-LTS
Benchmark                                                      (name)  Mode  Cnt    Score      Error   Units
UserAgentsBenchmark.parser                                    service  avgt    3    0.021 ±    0.002   us/op
UserAgentsBenchmark.parser:·gc.alloc.rate.norm                service  avgt    3   ≈ 10⁻⁵               B/op
UserAgentsBenchmark.parser                     valid-service-name-123  avgt    3    0.073 ±    0.005   us/op
UserAgentsBenchmark.parser:·gc.alloc.rate.norm valid-service-name-123  avgt    3   ≈ 10⁻⁵               B/op
UserAgentsBenchmark.regex                                     service  avgt    3    0.092 ±    0.055   us/op
UserAgentsBenchmark.regex:·gc.alloc.rate.norm                 service  avgt    3  128.000 ±    0.001    B/op
UserAgentsBenchmark.regex                      valid-service-name-123  avgt    3    0.299 ±    0.005   us/op
UserAgentsBenchmark.regex:·gc.alloc.rate.norm  valid-service-name-123  avgt    3  200.000 ±    0.001    B/op
```

```
JDK 17.0.4, OpenJDK 64-Bit Server VM, 17.0.4+8-LTS
Benchmark                                                      (name)  Mode  Cnt     Score      Error   Units
UserAgentsBenchmark.parser                                    service  avgt    3     0.015 ±    0.002   us/op
UserAgentsBenchmark.parser:·gc.alloc.rate.norm                service  avgt    3    ≈ 10⁻⁵               B/op
UserAgentsBenchmark.parser                     valid-service-name-123  avgt    3     0.030 ±    0.001   us/op
UserAgentsBenchmark.parser:·gc.alloc.rate.norm valid-service-name-123  avgt    3    ≈ 10⁻⁵               B/op
UserAgentsBenchmark.regex                                     service  avgt    3     0.078 ±    0.021   us/op
UserAgentsBenchmark.regex:·gc.alloc.rate.norm                 service  avgt    3   128.008 ±    0.043    B/op
UserAgentsBenchmark.regex                      valid-service-name-123  avgt    3     0.251 ±    0.017   us/op
UserAgentsBenchmark.regex:·gc.alloc.rate.norm  valid-service-name-123  avgt    3   200.013 ±    0.001    B/op
```


==COMMIT_MSG==
Optimize UserAgents

Optimize UserAgents.isValidName to use a hand-rolled implementation to validate user agent name to avoid
allocation and regex overhead.

Increase default UserAgents.format string builder buffer size and ensure capacity as we append user agents.
==COMMIT_MSG==

## Possible downsides?
Complexity of maintaining is higher than simple regex