@schlosna 2022-07-21
Tweaked the implementation a bit more exactly follow regex and improved a bit more:

```
JDK 17.0.4, OpenJDK 64-Bit Server VM, 17.0.4+8-LTS
Benchmark                                                      (name)  Mode  Cnt     Score     Error   Units
UserAgentsBenchmark.parser                                    service  avgt    3     0.008 ±   0.001   us/op
UserAgentsBenchmark.parser:·gc.alloc.rate.norm                service  avgt    3    ≈ 10⁻⁵              B/op
UserAgentsBenchmark.parser                     valid-service-name-123  avgt    3     0.018 ±   0.002   us/op
UserAgentsBenchmark.parser:·gc.alloc.rate.norm valid-service-name-123  avgt    3    ≈ 10⁻⁵              B/op
UserAgentsBenchmark.regex                                     service  avgt    3     0.077 ±   0.006   us/op
UserAgentsBenchmark.regex:·gc.alloc.rate.norm                 service  avgt    3   128.009 ±   0.028    B/op
UserAgentsBenchmark.regex                      valid-service-name-123  avgt    3     0.251 ±   0.148   us/op
UserAgentsBenchmark.regex:·gc.alloc.rate.norm  valid-service-name-123  avgt    3   200.013 ±   0.008    B/op
```