@svc-autorelease 2022-07-21
Released [2.27.0](https://github.com/palantir/conjure-java-runtime-api/releases/tag/2.27.0)